Renaming directories currently have an issue. If you rename (mv) a directory, its contents "disappear". To get the contents back, just re-create the directory.
This issue has led to more than one support issues.
This patch disallows renaming of directories and
returns a operation not supported message, e.g.

% mv dir1 dir2
mv: cannot move `dir1' to `dir2': Operation not supported

Another observation is that when a directory was renamed
it loses its directory status and becomes a normal file.

This is a first step to allow for supporting directory
renames. 



git-svn-id: http://s3fs.googlecode.com/svn/trunk@260 df820570-a93a-0410-bd06-b72b767a4274